Improve Python calling convention, remove early methods.
The new calling convention for Python functions has explicit keyword arguments
that consist of a Python string and a value of some sorts.  In other words,
Lisp keywords and Python keywords are now completely different things, thus
removing the semantic ambiguity of passing a Lisp keyword to a Python function.